platform_drv (x86): fix empty report when there are too many PCI devices
When there are too many PCI devices, the Expanding_reporter regenerates the report. However, this doesn't reset the BDF counter used to iterate over the devices. This results in starting the new report after the PCI device that triggered the report buffer overflow.
This commit fixes the issue by putting the BDF counter initialization inside the lambda function used to generate the report.

fixes genodelabs#3317